refactor: change file/folder name follow to conventions
