+++ src/components/Buttons/index.ts
@@ -1,0 +1,2 @@
+export * from './BasicButton/BasicButton'
+export * from './GradientButton/GradientButton'
