+++ src/components/Icons/TrystsLogoIcon/TrystLogoIcon.tsx
