+++ src/components/Icons/index.ts
@@ -1,0 +1,3 @@
+export * from './SearchIcon/SearchIcon'
+export * from './ShopIcon/ShopIcon'
+export * from './TrystsLogoIcon/TrystLogoIcon'
