--- src/components/button/index.ts
@@ -1,2 +1,0 @@
-export * from './basic-button/basic-button.component'
-export * from './gradient-button/gradient-button.component'
