+++ src/components/header/index.ts
@@ -1,2 +1,1 @@
-export * from './header.compnent'
-export * from './searchbar/searchbar.component'
+export * from './Header'
