--- src/components/icon/index.ts
@@ -1,3 +1,0 @@
-export * from './search-icon/search-icon.component'
-export * from './shop-icon/shop-icon.component'
-export * from './tryst-icon/trysts-icon.component'
